Major Project MRRC
This is a program for a car rental company called mates Rates rent a Car (MRRC) to help the operators of the company more efficiently manage their fleet of cars and their customers. For a complete comprehensive version of the program, run and build Part B as it contains all the functionality. Part A does not include search capabilities.